+++ priv/repo/migrations/20240407090614_add_locale_to_users.exs
@@ -1,0 +1,9 @@
+defmodule Claper.Repo.Migrations.AddTimezoneAndLocaleToUsers do
+  use Ecto.Migration
+
+  def change do
+    alter table(:users) do
+      add :locale, :string
+    end
+  end
+end